Move `ElementTable` into `table.rs`
